Now controls RGBW LEDs
Now controls RGBW LEDs but the response is very sluggish. The ATMega328 has only 2 hardware interrupts so the push-button is polled and currently it is often missed. The values for the white LED are added into the functions by bit-shifting which is processor intensive. I don't yet know a way around this, so I intend to modify the menu system to compensate by not updating the pixels on-the-fly…